requirements.txt
pyodbc
azure-identity